pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Modular Discord bot using Hikari framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hikari[speedups]>=2.0.0",
    "hikari-lightbulb>=2.3.5",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=23.0.0",
    "watchfiles>=0.20.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
music = [
    "wavelink>=3.0.0",
    "yt-dlp>=2023.7.6",
]

[project.scripts]
discord-bot = "bot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot", "config", "plugins"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["S101", "S311", "B008"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"plugins/*" = ["S110"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=bot --cov-report=term-missing"